Finished the Draculog_Executor Script
The Executor script is now in bug testing period, and it also doesn't have any test functions written yet. However, to do so I need to transform this script into a class as that would make for the easiest  test case Creation
