+++ Draculog_Downloader.py
@@ -1,70 +1,70 @@
-#### DracuLog - Downloader - Downloading Hook for Green Code and DracuLog Integration
-### Created by Daniel Jacoby alongside Dr. Joshua Gross, Aaron Helman, and Austin Folser
-##
-#
+### DracuLog   > Downloader
+## Abstract    - Managing script for all Downloaded Code from Various users and their submissions
+# Created by Daniel Jacoby alongside Dr. Joshua Gross, Aaron Helman, and Austin Folster
 
-# Notes:
-# Returned JSON from GreenCode looks as such:
-# JSON (Pulled Obj) = {
-#   PulledData (0) = {
-#       submissionId: int,
-#       userId: string,
-#       codeName: string like "dummy.cpp",
-#       mimetype: string like "cpp" or "c",
-#       codeString: string of entire code,
-#       codeCompile: int (0 or 1 for compiled or not)
-#   }, ...
-#   PulledData (N) = { ...
-#   }
-# }
-# Directory to store User code looks like:
-# Users_Code/
-#   User_ID/
-#       Submission_ID/
-#           Makefile
-#           JSON_Result (If tested)
-#           CPP_Code (Saved as userId_submissionId)
+"""
+Notes:
+Downloaded JSON from GreenCode looks as such:
+JSON (Pulled Obj) = {
+  PulledData [0] = {
+      submissionId: int,
+      userId: string,
+      codeName: string like "dummy.cpp",
+      mimetype: string like "cpp" or "c",
+      codeString: string of entire code,
+      codeCompile: int (0 or 1 for compiled or not)
+  }, ...
+  PulledData [N] = { ...
+  }
+}
 
+Directory to store User code looks like:
+Users_Code/
+  UserID/
+      SubmissionID/
+          Makefile
+          CPP_Code (Saved as userId_submissionId)
+"""
 
-# For Basic System Operations
+# String Used for if user's don't have code submitted to GreenCode
+
+# For Linux System Operations
+from datetime import datetime as dt
+import time
 import sys
 import os
 import shutil
 
-# For Downloading Data
-import json
-import requests
-from datetime import datetime as time
+# For Controlling the System
+from Draculog import GlobalValues as Globe
+from Draculog import SharedDraculogFunctions
+
+GreenCode = SharedDraculogFunctions()
 
 # Control Variables
-verbose = True
+verbose = Globe.verbose
+log = Globe.log
+control_file = None
 
-# API Variables
-greenCodeApiToken = ""
-greenCodeApiBase = "https://greencodemk2.herokuapp.com/code/"
-
-headers = {'Content-Type': 'application/json',
-           'Authorization': 'Bearer {0}'.format(greenCodeApiToken)}
+# List of Downloaded Code
+if not os.path.isfile(Globe.Newly_Downloaded_Code_Str):
+    Downloaded_File = open(Globe.Newly_Downloaded_Code_Str, "w+")
+else:
+    Downloaded_File = open(Globe.Newly_Downloaded_Code_Str, "a+")
 
 # Directory Variables
 mainCodeDirectory = "Downloaded_Code"
 
-### API Functions
-# Calls Green Code with a specific command for the Code route
-def Call_GreenCode_API(command):
-    apiCall = '{0}{1}'.format(greenCodeApiBase, command)
-    response = requests.get(apiCall, headers=headers)
-    if response.status_code == 200:
-        return json.loads(response.content.decode('utf-8'))
-    else:
-        return None
-
 ### Code Set Up Functions
 # Creates an MakeFile for the User's Code
-def Create_Makefile(path, codeFile):
-    # TODO Add Dynamic Compiler Variable
+def Create_Makefile(path, codeFile, User_Code_Exists):
+    # TODO-MultiCompiler This is what needs to be modified so that we can take in more than 1 compiler
+    if not User_Code_Exists:
+        GreenCode.Log_Time("UPDATE-*-\tUser code DNE, skipping", dt.now())
+        return
+    # TODO-MultiCompiler Add Dynamic Compiler Variable
     compiler = "gcc"
-    # TODO Add Dynamic Compiler Flags Variable
+    # TODO-MultiCompiler Add Dynamic Compiler Flags Variable
     flags = "-o"
     try:
         makefile = open(path + "/" + "Makefile", "w")
@@ -75,73 +75,105 @@ def Create_Makefile(path, codeFile):
         makefile.write("\t$(COM) $(FLAGS) code " + codeFile + "\n")
         makefile.close()
     except Exception as e:
-        print("Makefile Error -- " + e)
-        return False
-    return True
+        GreenCode.Log_Time("ERROR-*-\tMakefile creation error: " + str(e), dt.now(), Override=True)
+    return
+
 
 # Creates an Executable File from the User's Code
-def Create_User_Code(path, codeFile, codeString, testing=False):
+def Create_User_Code(path, codeFile, codeString):
     # Write the submitted code to the code file, and if it's None run a 60-second baseline measure
-    try:
-        file = open(path + "/" + codeFile, "w")
-        file.write(codeString if codeString is not None else "baseline 60")
-        file.close()
-    except Exception as e:
-        print("User Code Error -- " + e)
-        return False
-    if testing:
-        return True if codeString is not None else False
-    return True
+    file = open(path + "/" + codeFile, "w")
+    if codeString is not None:
+        file.write(codeString)
+    # TODO-MultiCompiler This is what needs to be modified so that we can take in more than 1 compiler
+    else:
+        GreenCode.Log_Time("ERROR-*-\tUser Code Doesn't Exist, Skipping", dt.now(), Override=True)
+
+    file.close()
+    return codeString is not None
+
+#
+def Add_User_Code_To_Wait_List(UserPath):
+    global Downloaded_File
+    if Downloaded_File is None:
+        Downloaded_File = open(Globe.Newly_Downloaded_Code_Str, "a+")
+    Downloaded_File.write(UserPath + "\n")
+    return
+
 
 # TODO Create a test where we pass it a list of dictionaries, passes if it makes them
 # Goes through all pulled users (from un-compiled code) and creates directories with a make file in them
 def Setup_UnCompiledCode(PulledJSON):
     # Make Main Directory
     if not os.path.isdir(mainCodeDirectory):
-        if verbose:
-            print("No main directory found, making one")
+        GreenCode.Log_Time("UPDATE-*-\tNo Main directory Found, Making new one", dt.now(), OnlyPrint=verbose)
         os.mkdir(mainCodeDirectory)
 
     for submission in PulledJSON:
         # Local Variables
-        u = str(submission["userId"])
-        s = str(submission["submissionId"])
-        m = str(submission["mimetype"])
-        codeFile = u + "_" + s + "." + m
+        u = str(submission["userId"])  # User ID
+        s = str(submission["submissionId"])  # Submission ID
+        m = str(submission["mimetype"])  # Mimetype (cpp/c/py etc)
+        codeFile = u + "_" + s + "." + m  # Code File Name
 
         # Checks if the User has a directory and if not, makes one
-        userpath = mainCodeDirectory + "/" + u + "_user"
+        userpath = mainCodeDirectory + "/" + u
         if not os.path.isdir(userpath):
-            if verbose:
-                print("Making a directory for user " + u)
+            GreenCode.Log_Time("UPDATE-*-\tNo User directory Found, Making new one", dt.now(), OnlyPrint=verbose)
             os.mkdir(userpath)
 
         # Checks if the User already has a submission of this number, and if not makes one
         submissionPath = userpath + "/" + s
         if os.path.isdir(submissionPath):
-            if verbose:
-                print("Submission already exists, removing previous submission")
+            GreenCode.Log_Time("UPDATE-*-\tSubmission already exists, removing previous submission", dt.now(), OnlyPrint=verbose)
             shutil.rmtree(submissionPath)
-        if verbose:
-            print("Making a new subdirectory for user " + u + "'s submission " + s)
+        GreenCode.Log_Time("UPDATE-*-\tMaking a new subdirectory for user "+u+"'s submission "+s, dt.now(), OnlyPrint=verbose)
+
         os.mkdir(submissionPath)
 
-        Create_Makefile(submissionPath, codeFile)
-        Create_User_Code(submissionPath, codeFile, submission["codeString"])
+        # TODO-MultiCompiler This is what needs to be modified so that we can take in more than 1 compiler
+        User_Code_Exists = Create_User_Code(submissionPath, codeFile, submission["codeString"])
+        Create_Makefile(submissionPath, codeFile, User_Code_Exists)
+
+        Add_User_Code_To_Wait_List(submissionPath)
 
     return
 
+
 ### Main Execution Build Section
 def main():
+    thisTime = dt.now()
+    GreenCode.Log_Time("DS##-\tDownload Starting", thisTime)
+
+    # Checks to see if we are already downloading, executing, or uploading code
+    global control_file
+    if os.path.isfile(Globe.Executing_Code_Str) or os.path.isfile(Globe.Uploading_Code_Str):
+        errorStr = "Executing" if os.path.isfile(Globe.Executing_Code_Str) else "Uploading"
+        GreenCode.Log_Time("WAIT-*-\tDraculog is currently " + errorStr + " code. . . . .", dt.now(), Override=True)
+        sys.exit(2)
+    if os.path.isfile(Globe.Downloading_Code_Str):
+        GreenCode.Log_Time("WAIT-*-\tDraculog is currently still Downloading code. . . . .", dt.now(), Override=True)
+        sys.exit(2)
+
     # Make a control Text File
-    tempFile = open("Downloading_Code.txt", "w")
-    UnCompiledCode = Call_GreenCode_API("notCompiled")
+    control_file = open(Globe.Downloading_Code_Str, "w")
+
+    # Download all Un-Compiled (ie Un-Tested) code from GreenCode's Website
+    # End Result is a File called "New_Downloaded_Code.txt" that contains all newly downloaded Submissions Paths
+    UnCompiledCode = GreenCode.Download_From_GreenCode()
+    if UnCompiledCode is None:
+        sys.exit(1)
     Setup_UnCompiledCode(UnCompiledCode)
+
     # Delete the control Text File
-    os.remove("Downloading_Code.txt")
+    os.remove(Globe.Downloading_Code_Str)
+
+    thisTime = dt.now()
+    GreenCode.Log_Time("DF##-\tDownload Finished", thisTime)
+
     return
 
 # Main Execution
 if __name__ == "__main__":
     main()
-    sys.exit()
+    sys.exit(0)
