+++ Sensors/Sensor_Load.py
@@ -2,28 +2,28 @@ import os
 import threading
 import time
 
-from Sensors.Sensor import GlobalSensorValues as Globe
-from Sensors.Sensor import Sensor
+from Sensor import GlobalSensorValues as Globe
 
-class Load(Sensor):
-    def __init__(self, interval, name):
-        super().__init__()
+class Load:
+    def __init__(self, name="Load", interval=Globe.interval, organizeMe=True):
         self.interval = interval
-        self.name = "Sensor " + name
+        self.name = "Sensor-" + name
         self.thread = None
         self.data = []
         self.success = 0
         self.failure = 0
-
+        self.load = 0.0
+        self.organizeMe = organizeMe
         return
 
     def callMe(self):
-        print("Hi, I'm " + self.name)
+        print("Hi, I'm " + self.name + " running at " + self.interval)
         return
 
     def Build_Logger(self, function=None):
         if function is None:
-            function = self.Log_Test
+            print("ERROR-*-\tNo function argument was passed, returning without building a logger.")
+            return
         self.thread = threading.Thread(target=function, name=self.name)
         return
 
@@ -32,31 +32,30 @@ class Load(Sensor):
         return
 
     def Log(self):
-        load = 0.0
+        self.load = 0.0
         while Globe.continueLogging:
             try:
-                load = os.getloadavg()[0]
+                self.load = os.getloadavg()[0]
             except IndexError or RuntimeError:
                 self.failure += 1
-                load = 0.0
+                self.load = 0.0
                 pass
             this_time = time.time()
-            if load is not None:
+            if self.load is not None:
                 self.success += 1
             else:
                 self.failure += 1
-                load = 0.0
+                self.load = 0.0
 
-            self.data.append((float(this_time), load))
+            self.data.append((float(this_time), self.load))
             time.sleep(self.interval)
         return
 
-    def Log_Test(self):
-        super().Log_Test()
-        return
-
     def End_Logging(self):
-        return
+        data_copy = self.data.copy()
+        self.data.clear()
+        self.thread.join()
+        return data_copy
 
     def Get_Data(self):
         return self.data
