+++ Sensors/Sensor_PyRAPL.py
@@ -2,24 +2,25 @@
 
 import pyRAPL
 
-from Sensors.Sensor import GlobalSensorValues as Globe
-from Sensors.Sensor import Sensor
+from Sensor import GlobalSensorValues as Globe
 
-class PyRAPL(Sensor):
-    def __init__(self, interval, name):
+class PyRAPL:
+    def __init__(self, name="PyRAPL", interval=Globe.interval, organizeMe=True):
         super().__init__()
         self.interval = interval
-        self.name = "Sensor " + name
+        self.name = "Sensor-" + name
         self.meter = None
         self.data = []
+        pyRAPL.setup(devices=[pyRAPL.Device.PKG])
+        self.organizeMe = organizeMe
         return
 
     def CallMe(self):
-        print("Hi, I'm " + self.name)
+        print("Hi, I'm " + self.name + " running at " + self.interval)
         return
 
-    def Build_Logger(self, sName):
-        pyRAPL.setup(devices=[pyRAPL.Device.PKG])
+    def Build_Logger(self, function):
+        sName = function()
         self.meter = pyRAPL.Measurement(sName)
         return
 
@@ -27,22 +28,23 @@ class PyRAPL(Sensor):
         self.meter.begin()
         return
 
-    def Log_Test(self):
-        super().Log_Test()
-        return
+    def Log(self):
+        return self.name
 
     def Compile_Energy_Data(self):
-        self.data.append((float(self.meter.timestamp), self.meter.result))
+        self.data.append(self.meter.result)
         return
 
     def End_Logging(self):
         self.meter.end()
-        del self.meter
+        self.Compile_Energy_Data()
         self.meter = None
-        return
+        data_copy = self.data.copy()
+        self.data.clear()
+        return data_copy
 
     def Get_Data(self):
-        return self.data
+        return self.data[0]
 
     def Print_Data(self):
         print("==========Energy==========")
