+++ Sensors/Sensor_Time.py
@@ -3,27 +3,27 @@
 import threading
 import time
 
-from Sensors.Sensor import GlobalSensorValues as Globe
-from Sensors.Sensor import Sensor
+from Sensor import GlobalSensorValues as Globe
 
-class Time(Sensor):
-    def __init__(self, interval, name):
-        super().__init__()
+class Time:
+    def __init__(self, name="Time", interval=Globe.interval, organizeMe=True):
         self.interval = interval
-        self.name = "Sensor " + name
+        self.name = "Sensor-" + name
         self.thread = None
         self.data = []
         self.success = 0
         self.failure = 0
+        self.organizeMe = organizeMe
         return
 
     def Call_Me(self):
-        print("Hi, I'm " + self.name)
+        print("Hi, I'm " + self.name + " running at " + self.interval)
         return
 
     def Build_Logger(self, function=None):
         if function is None:
-            function = self.Log_Test
+            print("error: No function argument was passed, returning without building a logger.")
+            return
         self.thread = threading.Thread(target=function, name=self.name)
         return
 
@@ -37,12 +37,11 @@ class Time(Sensor):
             time.sleep(self.interval)
         return
 
-    def Log_Test(self):
-        super().Log_Test()
-        return
-
     def End_Logging(self):
-        return
+        data_copy = self.data.copy()
+        self.data.clear()
+        self.thread.join()
+        return data_copy
 
     def Get_Data(self):
         return self.data
