--- raw_data_csv.csv
@@ -1,29 +1,0 @@
-Run Number:,0,Parameters:,200000 b,Executable:,sort,Results File:,Source/sorts/200000 b_results.txt,Baseline CPU Temp:,60,Total Energy (Micro Joules):,2197137902.0
-Script Delta:,140.78404784202576,Script Start:,1641781539.5314558,Script End:,1641781680.3155036
- 
-time,cpu,load
-1641781539.529716,27.0,0.0
-1641781545.5357487,36.0,0.08
-1641781551.541778,37.0,0.15
-1641781557.5478039,38.0,0.28
-1641781563.553821,39.0,0.34
-1641781569.5598392,40.0,0.39
-1641781575.5604985,39.0,0.44
-1641781581.5665157,40.0,0.49
-1641781587.5725327,41.0,0.57
-1641781593.5785506,40.0,0.6
-1641781599.5845675,40.0,0.63
-1641781605.5905848,40.0,0.66
-1641781611.5966022,40.0,0.69
-1641781617.6026177,41.0,0.74
-1641781623.6086352,41.0,0.76
-1641781629.6146524,42.0,0.78
-1641781635.6206703,43.0,0.8
-1641781641.6266875,43.0,0.81
-1641781647.6327035,44.0,0.84
-1641781653.6387205,43.0,0.86
-1641781659.6447372,43.0,0.87
-1641781665.6507566,44.0,0.88
-1641781671.6567736,44.0,0.89
-1641781677.6627915,43.0,0.91
-
