--- raw_data_json.json
@@ -1,1 +1,0 @@
-{"Run Number": "0", "Parameters": "200000 b", "Executable": "sort", "Results File": "Source/sorts/200000 b_results.txt", "Baseline CPU Temp": 60, "Script Delta": 140.78404784202576, "Script Start": 1641781539.5314558, "Script End": 1641781680.3155036, "time": [1641781539.529716, 1641781545.5357487, 1641781551.541778, 1641781557.5478039, 1641781563.553821, 1641781569.5598392, 1641781575.5604985, 1641781581.5665157, 1641781587.5725327, 1641781593.5785506, 1641781599.5845675, 1641781605.5905848, 1641781611.5966022, 1641781617.6026177, 1641781623.6086352, 1641781629.6146524, 1641781635.6206703, 1641781641.6266875, 1641781647.6327035, 1641781653.6387205, 1641781659.6447372, 1641781665.6507566, 1641781671.6567736, 1641781677.6627915], "cpu": [[1641781539.538879, 27.0], [1641781545.5466337, 36.0], [1641781551.554138, 37.0], [1641781557.5581346, 38.0], [1641781563.5621214, 39.0], [1641781569.5661123, 40.0], [1641781575.5701292, 39.0], [1641781581.5737424, 40.0], [1641781587.5814347, 41.0], [1641781593.5891266, 40.0], [1641781599.596813, 40.0], [1641781605.604464, 40.0], [1641781611.6121013, 40.0], [1641781617.6197352, 41.0], [1641781623.6273746, 41.0], [1641781629.6350098, 42.0], [1641781635.642654, 43.0], [1641781641.650288, 43.0], [1641781647.6579683, 44.0], [1641781653.6656353, 43.0], [1641781659.6732986, 43.0], [1641781665.6781476, 44.0], [1641781671.6858182, 44.0], [1641781677.6935, 43.0]], "load": [[1641781539.530826, 0.0], [1641781545.536525, 0.08], [1641781551.5418043, 0.15], [1641781557.5478551, 0.28], [1641781563.5538893, 0.34], [1641781569.559923, 0.39], [1641781575.5659564, 0.44], [1641781581.572, 0.49], [1641781587.5780365, 0.57], [1641781593.5840704, 0.6], [1641781599.5901039, 0.63], [1641781605.596137, 0.66], [1641781611.6021702, 0.69], [1641781617.6082032, 0.74], [1641781623.614236, 0.76], [1641781629.6202698, 0.78], [1641781635.6263032, 0.8], [1641781641.6323369, 0.81], [1641781647.6365125, 0.84], [1641781653.6425545, 0.86], [1641781659.6485877, 0.87], [1641781665.6546211, 0.88], [1641781671.6606555, 0.89], [1641781677.6666899, 0.91]], "pyrapl": [], "Total Energy (Micro Joules)": 2197137902.0},
